moved default functionality from broker impl to trait
